+++ alacritty/src/renderer/mod.rs
@@ -9,7 +9,7 @@ use ahash::RandomState;
 use crossfont::Metrics;
 use glutin::context::{ContextApi, GlContext, PossiblyCurrentContext};
 use glutin::display::{GetGlDisplay, GlDisplay};
-use log::{debug, error, info, warn, LevelFilter};
+use log::{debug, info, LevelFilter};
 use unicode_width::UnicodeWidthChar;
 
 use alacritty_terminal::graphics::UpdateQueues;
@@ -101,6 +101,7 @@ pub struct Renderer {
     text_renderer: TextRendererProvider,
     rect_renderer: RectRenderer,
     graphics_renderer: GraphicsRenderer,
+    robustness: bool,
 }
 
 /// Wrapper around gl::GetString with error checking and reporting.
@@ -148,6 +149,9 @@ impl Renderer {
         info!("Running on {renderer}");
         info!("OpenGL version {gl_version}, shader_version {shader_version}");
 
+        // Check if robustness is supported.
+        let robustness = Self::supports_robustness();
+
         let is_gles_context = matches!(context.context_api(), ContextApi::Gles(_));
 
         // Use the config option to enforce a particular renderer configuration.
@@ -181,7 +185,7 @@ impl Renderer {
             }
         }
 
-        Ok(Self { text_renderer, rect_renderer, graphics_renderer })
+        Ok(Self { text_renderer, rect_renderer, graphics_renderer, robustness })
     }
 
     pub fn draw_cells<I: Iterator<Item = RenderableCell>>(
@@ -212,10 +216,10 @@ impl Renderer {
         glyph_cache: &mut GlyphCache,
     ) {
         let mut wide_char_spacer = false;
-        let cells = string_chars.enumerate().map(|(i, character)| {
+        let cells = string_chars.enumerate().filter_map(|(i, character)| {
             let flags = if wide_char_spacer {
                 wide_char_spacer = false;
-                Flags::WIDE_CHAR_SPACER
+                return None;
             } else if character.width() == Some(2) {
                 // The spacer is always following the wide char.
                 wide_char_spacer = true;
@@ -224,7 +228,7 @@ impl Renderer {
                 Flags::empty()
             };
 
-            RenderableCell {
+            Some(RenderableCell {
                 point: Point::new(point.line, point.column + i),
                 character,
                 extra: None,
@@ -233,7 +237,7 @@ impl Renderer {
                 fg,
                 bg,
                 underline: fg,
-            }
+            })
         });
 
         self.draw_cells(size_info, glyph_cache, cells);
@@ -284,6 +288,49 @@ impl Renderer {
                 alpha,
             );
             gl::Clear(gl::COLOR_BUFFER_BIT);
+        }
+    }
+
+    /// Get the context reset status.
+    pub fn was_context_reset(&self) -> bool {
+        // If robustness is not supported, don't use its functions.
+        if !self.robustness {
+            return false;
+        }
+
+        let status = unsafe { gl::GetGraphicsResetStatus() };
+        if status == gl::NO_ERROR {
+            false
+        } else {
+            let reason = match status {
+                gl::GUILTY_CONTEXT_RESET_KHR => "guilty",
+                gl::INNOCENT_CONTEXT_RESET_KHR => "innocent",
+                gl::UNKNOWN_CONTEXT_RESET_KHR => "unknown",
+                _ => "invalid",
+            };
+
+            info!("GPU reset ({})", reason);
+
+            true
+        }
+    }
+
+    fn supports_robustness() -> bool {
+        let mut notification_strategy = 0;
+        if GlExtensions::contains("GL_KHR_robustness") {
+            unsafe {
+                gl::GetIntegerv(gl::RESET_NOTIFICATION_STRATEGY_KHR, &mut notification_strategy);
+            }
+        } else {
+            notification_strategy = gl::NO_RESET_NOTIFICATION_KHR as gl::types::GLint;
+        }
+
+        if notification_strategy == gl::LOSE_CONTEXT_ON_RESET_KHR as gl::types::GLint {
+            info!("GPU reset notifications are enabled");
+            true
+        } else {
+            info!("GPU reset notifications are disabled");
+            false
         }
     }
 
@@ -387,7 +434,7 @@ impl GlExtensions {
 
 extern "system" fn gl_debug_log(
     _: gl::types::GLenum,
-    kind: gl::types::GLenum,
+    _: gl::types::GLenum,
     _: gl::types::GLuint,
     _: gl::types::GLenum,
     _: gl::types::GLsizei,
@@ -395,11 +442,5 @@ extern "system" fn gl_debug_log(
     _: *mut std::os::raw::c_void,
 ) {
     let msg = unsafe { CStr::from_ptr(msg).to_string_lossy() };
-    match kind {
-        gl::DEBUG_TYPE_ERROR | gl::DEBUG_TYPE_UNDEFINED_BEHAVIOR => {
-            error!("[gl_render] {}", msg)
-        },
-        gl::DEBUG_TYPE_DEPRECATED_BEHAVIOR => warn!("[gl_render] {}", msg),
-        _ => debug!("[gl_render] {}", msg),
-    }
+    debug!("[gl_render] {}", msg);
 }