+++ alacritty_terminal/CHANGELOG.md
@@ -8,7 +8,19 @@ sections should follow the order `Added`, `Changed`, `Deprecated`, `Fixed` and
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## 0.24.2-dev
+## 0.25.1-dev
+
+### Changed
+
+- Pass `-q` to `login` on macOS if `~/.hushlogin` is present
+
+## 0.25.0
+
+### Changed
+
+- Replaced `Options::hold` with `Options::drain_on_exit`
+
+## 0.24.2
 
 ### Added